run_x_to_nwb_conversion: Add mode for creating meta data files

+++ ipfx/bin/run_x_to_nwb_conversion.py
@@ -7,7 +7,7 @@ from ipfx.x_to_nwb.ABFConverter import ABFConverter
 from ipfx.x_to_nwb.DatConverter import DatConverter
 
 
-def convert(inFileOrFolder, overwrite=False, fileType=None):
+def convert(inFileOrFolder, overwrite=False, fileType=None, outputMetadata=False):
     """
     Convert the given file to a NeuroDataWithoutBorders file using pynwb
 
@@ -18,6 +18,7 @@ def convert(inFileOrFolder, overwrite=False, fileType=None):
     :param inFileOrFolder: path to a file or folder
     :param overwrite: overwrite output file, defaults to `False`
     :param fileType: file type to be converted, must be passed iff `inFileOrFolder` refers to a folder
+    :param outputMetadata: output metadata of the file, helpful for debugging
 
     :return: path of the created NWB file
     """
@@ -40,16 +41,23 @@ def convert(inFileOrFolder, overwrite=False, fileType=None):
 
     outFile = root + ".nwb"
 
-    if os.path.exists(outFile):
+    if not outputMetadata and os.path.exists(outFile):
         if overwrite:
             os.remove(outFile)
         else:
             raise ValueError(f"The output file {outFile} does already exist.")
 
     if ext == ".abf":
-        ABFConverter(inFileOrFolder, outFile)
+        if outputMetadata:
+            ABFConverter.outputMetadata(inFileOrFolder)
+        else:
+            ABFConverter(inFileOrFolder, outFile)
     elif ext == ".dat":
-        DatConverter(inFileOrFolder, outFile)
+        if outputMetadata:
+            DatConverter.outputMetadata(inFileOrFolder)
+        else:
+            DatConverter(inFileOrFolder, outFile)
+
     else:
         raise ValueError(f"The extension {ext} is currently not supported.")
 
@@ -67,6 +75,8 @@ def main():
     parser.add_argument("--fileType", type=str, default=None, choices=[".abf"],
                         help=("Type of the files to convert (only required "
                               "if passing folders)."))
+    parser.add_argument("--outputMetadata", action="store_true", default=False,
+                        help="Helper for debugging which outputs HTML/TXT files with the metadata contents of the files.")
     parser.add_argument("filesOrFolders", nargs="+",
                         help="List of ABF files/folders to convert.")
 
@@ -80,7 +90,8 @@ def main():
 
     for fileOrFolder in args.filesOrFolders:
         print(f"Converting {fileOrFolder}")
-        convert(fileOrFolder, overwrite=args.overwrite, fileType=args.fileType)
+        convert(fileOrFolder, overwrite=args.overwrite, fileType=args.fileType,
+                outputMetadata=args.outputMetadata)
 
 
 if __name__ == "__main__":
